battery_supply_agent/pyproject.toml
[project]
name = "battery_supply_agent"
version = "0.1.0"
description = "A standalone agent for battery supply"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-adk>=1.8.0",
    "fastapi>=0.111.0",
    "uvicorn>=0.30.1",
    "python-dotenv>=1.0.0",
    "httpx>=0.28.1",
    "a2a-sdk>=0.2.16",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
py-modules = ["agent", "agent_executor"]
